Updates to genNixpacks.
When nixpacks is run, we also update the file names in Docker manually
from .nixpacks to .astral/.nixpacks.
This keeps us from pushing a Dockerfile to the root of the repo it's
being run in.
Keeps things clean, too.

This is better (IMO) than forking the nixpacks repo, and is an easy
RegEx replace.